refactor: HeatMapImageBox 별도 파일로 분리
