+++ app/user-web-prototype/src/components/presentational/model-learning/data-navigation/HeatMapImageBox.jsx
@@ -1,0 +1,24 @@
+import React from "react";
+import Box from "@mui/material/Box";
+import CardMedia from "@mui/material/CardMedia";
+
+const HeatMapImageBox = (props: { imageUrl: string }) => {
+  return (
+    <Box sx={{
+      display: "flex",
+      justifyContent: "center",
+      marginTop: "40px"
+    }}>
+      <CardMedia
+        component="img"
+        image={props.imageUrl}
+        alt="heatmap image"
+        sx={{
+          width: "50%"
+        }}
+      />
+    </Box>
+  );
+}
+
+export default HeatMapImageBox;
